+++ src/systems/mod.rs
@@ -1,5 +1,5 @@
 mod move_pieces;
 pub use self::move_pieces::MovePiecesSystem;
 
-mod mouse;
-pub use self::mouse::MouseRaycastSystem;
+mod mouse_raycast;
+pub use self::mouse_raycast::MouseRaycastSystem;
